agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = "Starter"
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
license = "MIT"

[project]
name = "agent"
version = "0.0.1"
dependencies = [
    "langchain-openai",
    "langchain-anthropic",
    "langchain",
    "openai",
    "langchain-community",
    "copilotkit",
    "uvicorn",
    "python-dotenv",
    "langchain-core",
    "langgraph-cli",
    "requests (>=2.32.3,<3.0.0)",
    "ragflow-sdk (>=0.17.1,<0.18.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "chromadb (>=0.6.3,<0.7.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "chroma (>=0.2.0,<0.3.0)"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain-openai = "^0.2.1"
langchain-anthropic = "^0.2.1"
langchain = "^0.3.1"
openai = "^1.51.0"
langchain-community = "^0.3.1"
copilotkit = "0.1.39"
uvicorn = "^0.31.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}

[tool.poetry.scripts]
demo = "agent.demo:main"
test = "agent.test:test"
test_sdk = "agent.test_rag_python_sdk:main"

[tool.setuptools.packages.find]
exclude = ["chroma_store"]